requirements.txt
pandas
django-admin
django-crispy-forms
django-pandas
